split read data functions to separate module
